Add files via upload
I changed a few things....made a single Thomas algorithm function that takes in the diagonal elements and the rhs, it takes in vectors, even though that means a vector of constants.

I thought I had all the errors gone before but some were still there for part B.

Part A is Solid. which makes me think its the F function....but I don't see what else to do with it.

I just cant get part B as accurate right now....SO there is probably some error with filling the rhs input vector.

But this isn't actually due until the 19th so I got plenty of time.