Add files via upload
Deleted some comments, and deleted some unecessarily lines I had included just  for included for consistency between part a and part b. I also had a spacing issue for the last input to the rhs vector that i didnt notice before. The error didnt show up becuase the values were so close anyway that I didnt notice. Everything should be good now. 

Thoughts for the grid convergene: My values are already so close as to be indistinguishable from the exact solution without zooming in a whole bunch. Average error is on the order of 1E-3...And total error is proportional to dt^2 and dx^2....I think I will just try to get the error below 0.001 for both problems. And I will just do it manually by changing the inputs, I am not gonna do a while loop like last time.